Move mappings to constants.py
