+++ src/pocketchemist_nmr/spectra/nmrpipe/constants.py
@@ -2,6 +2,9 @@
 Constants for NMRPipe spectra
 """
 from enum import Enum
+import typing as t
+
+from ..constants import DomainType
 
 __all__ = ('header_size_bytes', 'data_size_bytes', 'Plane2DPhase',
            'SignAdjustment')
@@ -32,3 +35,52 @@ class SignAdjustment(Enum):
     NEGATE_IMAG = 16.0  # Negate the imaginary component
     REAL_NEGATE_IMAG = 17.0  # Same as REAL + NEGATE_IMAG
     COMPLEX_NEGATE_IMAG = 18.0  # Same as COMPLEX + NEGATE_IMAG
+
+
+# Mappings between enum values and NMRPipe header values
+mappings = {
+    'domain_type': {0.0: DomainType.TIME,
+                    1.0: DomainType.FREQ,
+                    None: DomainType.UNKNOWN},
+    'sign_adjustment': {1.0: SignAdjustment.REAL,
+                        2.0: SignAdjustment.COMPLEX,
+                        16.0: SignAdjustment.NEGATE_IMAG,
+                        17.0: SignAdjustment.REAL_NEGATE_IMAG,
+                        18.0: SignAdjustment.COMPLEX_NEGATE_IMAG,
+                        None: SignAdjustment.NONE},
+    'fd2dphase': {0.0: Plane2DPhase.MAGNITUDE,
+                  1.0: Plane2DPhase.TPPI,
+                  2.0: Plane2DPhase.STATES,
+                  3.0: Plane2DPhase.IMAGE,
+                  4.0: Plane2DPhase.ARRAY,
+                  None: Plane2DPhase.NONE},
+}
+
+
+def find_mapping(name, cnst, reverse=False, round_cnst=True) \
+        -> t.Union[DomainType, SignAdjustment, Plane2DPhase]:
+    """Find the mapping for constant (enum) values.
+
+    Parameters
+    ----------
+    name
+        The name of the mapping to use. e.g. 'domain_type'
+    cnst
+        The cnst to retrieve for the mapping
+    reverse
+        If False (default), map the cnst to the mapping dict key
+        If True, map the cnst to the mapping dict value
+    round_cnst
+        If the cnst is a floating point number, round it to the nearest
+        integer float value. (ex: 3.9 -> 4.0)
+    """
+    d_mapping = mappings[name]
+    none_value = d_mapping[None]
+    if reverse:
+        d_mapping = {v: k for k, v in d_mapping.items()}
+        none_value = None
+
+    # Clean the cnst, if needed
+    if round_cnst and isinstance(cnst, float):
+        cnst = round(cnst, 1)
+    return d_mapping.get(cnst, none_value)
